Show messages when control submitted or touched
Note that the implementation notifies each ValidationMessagesComponent
separately, instead of storing the Form submitted state once in a
parent component. Reason being that a form which is submitted and then
changed afterwards might show additional form controls (due to
different answers). The validation messages of these controls shouldn't
be shown until the form is submitted again (or they are touched).